Vision-Language-Model-based-AI-for-Interactive-Universal-Robotic-Manipulation_softwareCV/requirements.txt
# Core Data Science & Image Processing
numpy==1.26.4
opencv-python==4.9.0.80
pillow==10.3.0
matplotlib==3.8.4
pandas==2.2.2  # Explicitly added for ultralytics dependency
scikit-image==0.23.2

# Machine Learning & Deep Learning (PyTorch for macOS arm64)
# Ensure you have the correct PyTorch version for your specific HW/OS if not macOS arm64
# Check: https://pytorch.org/get-started/locally/
torch==2.2.0
torchvision==0.17.0

# LLM & NLP Libraries
transformers==4.40.1
openai==1.33.0
google-generativeai==0.5.2
tiktoken==0.7.0  # Required by OpenAI
httpx==0.27.0   # Explicitly specify compatible version for openai

# Computer Vision & Object Detection
ultralytics==8.2.2  # YOLOv8
mediapipe==0.10.14 # For Hand Gesture Recognition

# Audio Processing
pyaudio==0.2.14
SpeechRecognition==3.10.1

# Utilities & Configuration
pydantic==1.10.15 # Using latest V1 for compatibility with existing code
jsonschema==4.22.0
pyyaml==6.0.1
python-dotenv==1.0.1
requests==2.31.0 # For API calls
retry==0.9.2
typing-extensions==4.11.0 # Often required by newer libraries
